Move delete workplace component
