+++ frontend/src/app/features/subsidiary/subsidiary-routing.module.ts
@@ -67,7 +67,7 @@ import { ViewSubsidiaryStaffRecordsComponent } from './staff-records/view-subsid
 import { ViewSubsidiaryTrainingAndQualificationsComponent } from './training-and-qualifications/view-subsidiary-training-and-qualifications.component';
 import { ViewSubsidiaryWorkplaceUsersComponent } from './workplace-users/view-subsidiary-workplace-users.component';
 import { ViewSubsidiaryWorkplaceComponent } from './workplace/view-subsidiary-workplace.component';
-import { DeleteWorkplaceComponent } from './delete-workplace/delete-workplace.component';
+import { DeleteWorkplaceComponent } from '../workplace/delete-workplace/delete-workplace.component';
 
 // eslint-disable-next-line max-len
 const routes: Routes = [
@@ -163,11 +163,6 @@ const routes: Routes = [
         component: ViewSubsidiaryWorkplaceComponent,
         data: { title: 'Workplace' },
       },
-      {
-        path: 'delete-workplace',
-        component: DeleteWorkplaceComponent,
-        data: { title: 'Delete workplace' },
-      },
     ],
   },
   {
@@ -237,7 +232,11 @@ const routes: Routes = [
           title: 'Regulated by CQC',
         },
       },
-
+      {
+        path: 'delete-workplace',
+        component: DeleteWorkplaceComponent,
+        data: { title: 'Delete workplace' },
+      },
       {
         path: 'select-workplace',
         component: SelectWorkplaceComponent,
